Added basic collision to block and barrier.
Note: Problem ensues when velocity is too great and block passes through. May need to create an invisible barrier that pushes block inside bounds. Or maybe check when the block is outside of rectangle to pull it back in. Oh that sounds good. Let's do that internal thoughts....

Let's hope no one reads this.